+++ cores/maple/libmaple/util.h
@@ -86,7 +86,7 @@ extern "C"{
 #define __swap16(u16)  __builtin_bswap16(u16)  ///< built-in function to swap Endian of 16-bit number
 
 #define memclr(buffer, size)  memset(buffer, 0, size)
-#define varclr(_var)          memclr(&_var, sizeof(_var))
+#define varclr(_var)          memclr(_var, sizeof(*(_var)))
 #define arrcount(_arr)       ( sizeof(_arr) / sizeof(_arr[0]) )
 
 static inline int stringlen(const char* str)