+++ changelog.md
@@ -2,6 +2,9 @@
 
 - setNtpServer
 - Close #69 Write feather_dfu using C + libusb instead of using python due to version conflict/confusion
+- Add IP Address to FeatherAP SoftAP's join/leave callback
+	- change signature (const uint8_t mac[6]) --> (const uint8_t mac[6], uint32_t ipv4)
+	- Add clientIP(id) to FeatherAP
 
 # 0.6.2